requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
python-dotenv
jwt
pydantic
pydantic[email]
apscheduler